feat(v2): supervisord controlled container
This change makes raspap-docker use supervisord instead of systemd.
The current version is tested to be functional with "out of the box connectivity".
More work is needed to support process restarts from config changes (possibly in raspap-webgui).
The docker container uses the raspap-ansible repo with a slightly modified playbook.
There are a few more outstanding issues that need addressed:

1. dhcpcd is not playing nicely inside the container. To get a static ip on wlan0 I had to put the RaspAP config outside of the container on the host's dhcpcd.
2. iptable rules need to be handled at 'docker run' time, not at docker build time (where they are excluded)
3. Configuration should be able to persist if a new version of the raspap docker image comes out. Basically let's support volume mounts
4. A good way to depend on raspap-ansible. Is a submodule overkill? Good enough to git clone inside the docker build?

This PR is a work-in-progress.